refactor #101 Mapper, Remote 디렉토리 제거
