+++ lib/features/auth/data/data_sources/auth_reissue_data_source.dart
@@ -1,5 +1,5 @@
-import 'package:debateseason_frontend_v1/features/auth/data/models/remote/request/auth_reissue_req.dart';
-import 'package:debateseason_frontend_v1/features/auth/data/models/remote/response/auth_reissue_res.dart';
+import 'package:debateseason_frontend_v1/features/auth/data/models/request/auth_reissue_req.dart';
+import 'package:debateseason_frontend_v1/features/auth/data/models/response/auth_reissue_res.dart';
 import 'package:debateseason_frontend_v1/utils/base/base_res.dart';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
