+++ lib/features/auth/data/models/request/users_login_req.dart
@@ -1,3 +1,4 @@
+import 'package:debateseason_frontend_v1/features/auth/domain/entities/users_login_entity.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'users_login_req.g.dart';
@@ -15,4 +16,11 @@ class UsersLoginReq {
   toJson() => _$UsersLoginReqToJson(this);
 
   fromJson(json) => _$UsersLoginReqFromJson(json);
+
+  factory UsersLoginReq.fromEntity(UsersLoginEntity entity) {
+    return UsersLoginReq(
+      identifier: entity.identifier,
+      socialType: entity.socialType,
+    );
+  }
 }
