+++ lib/features/auth/data/models/response/users_login_res.dart
@@ -1,3 +1,4 @@
+import 'package:debateseason_frontend_v1/features/auth/domain/entities/users_login_entity.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'users_login_res.g.dart';
@@ -20,4 +21,11 @@ class UsersLoginRes {
       _$UsersLoginResFromJson(json);
 
   Map<String, dynamic> toJson() => _$UsersLoginResToJson(this);
+
+  toEntity() => UsersLoginEntity(
+        accessToken: accessToken,
+        refreshToken: refreshToken,
+        socialType: socialType,
+        profileStatus: profileStatus,
+      );
 }
