+++ lib/features/auth/data/repository_impls/auth_reissue_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:debateseason_frontend_v1/core/services/secure_storage_service.dart';
-import 'package:debateseason_frontend_v1/features/auth/data/data_sources/remote/auth_reissue_data_source.dart';
-import 'package:debateseason_frontend_v1/features/auth/data/mapper/auth_reissue_mapper.dart';
-import 'package:debateseason_frontend_v1/features/auth/domain/repositories/remote/auth_reissue_repository.dart';
+import 'package:debateseason_frontend_v1/features/auth/data/data_sources/auth_reissue_data_source.dart';
+import 'package:debateseason_frontend_v1/features/auth/data/models/request/auth_reissue_req.dart';
+import 'package:debateseason_frontend_v1/features/auth/domain/repositories/auth_reissue_repository.dart';
 
 class AuthReissueRepositoryImpl implements AuthReissueRepository {
   final AuthReissueDataSource dataSource;
@@ -15,7 +15,7 @@ class AuthReissueRepositoryImpl implements AuthReissueRepository {
     final SecureStorageService storage = SecureStorageService();
 
     final response = await dataSource.postAuthReissue(
-      body: AuthReissueMapper.toRequest(refreshToken: refreshToken),
+      body: AuthReissueReq(refreshToken: refreshToken),
     );
 
     await Future.wait([
