+++ lib/features/auth/presentation/view_models/auth_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:debateseason_frontend_v1/core/services/shared_preferences_service.dart';
 import 'package:debateseason_frontend_v1/features/auth/auth_constants.dart';
 import 'package:debateseason_frontend_v1/features/auth/domain/entities/users_login_entity.dart';
-import 'package:debateseason_frontend_v1/features/auth/domain/repositories/remote/users_login_repository.dart';
+import 'package:debateseason_frontend_v1/features/auth/domain/repositories/users_login_repository.dart';
 import 'package:debateseason_frontend_v1/features/profile/domain/repositories/remote/profile_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 import 'package:debateseason_frontend_v1/utils/logger.dart';
